requirements.txt
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environs==10.3.0
eskiz-sms==0.2.3
exceptiongroup==1.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
